requirements/base.txt
Django==2.2
psycopg2-binary==2.8.5
rules==2.0.1
Pillow==8.1.2
chardet==3.0.4
django_select2==7.2.4
django-simple-history==2.12.0
django-hashid-field==3.2.1
